Wave 2 — Core C++ refactor (src F-1, F-3, F-4, F-5, F-6, F-7, F-8, F-9, F-10, F-11, F-14, F-15, F-16, F-17, F-18, F-19, F-21, F-22, F-24)
- Header: Standard profile defaults via KdfProfiles, const buffer(), size_t for nonce/tag sizes
- FileTable: FileEntry uint64_t fields, next_write_offset recovery, O(N) duplicate-name suffixing
- FileManager: file-table read inside slot recovery loop, full-file-path init(), nextAvailableContainerPath helper, capacity-check dedup, magic pre-check removal, stronger file_table_size validation, setKdfParams clamping
- EncryptPipeline: clarified push-or-EOF short-circuit
- DecryptPipeline: removed double skipSlots(entry.offset); std::filesystem path concat

+++ include/FileTable.h
@@ -1,6 +1,7 @@
 #ifndef FILE_TABLE_H
 #define FILE_TABLE_H
 
+#include <cstdint>
 #include <string>
 #include <vector>
 
@@ -18,9 +19,9 @@
 
 struct FileEntry {
     std::string name;
-    size_t size;
-    size_t offset;
-    size_t chunks;
+    uint64_t size;
+    uint64_t offset;
+    uint64_t chunks;
     std::string checksum_sha256;
 };
 
@@ -29,8 +30,8 @@ public:
     FileTable();
     ~FileTable();
 
-    void addFileEntry(const std::string& pathToFile, const std::string& checkSum, size_t offset,
-                      size_t actual_size);
+    void addFileEntry(const std::string& pathToFile, const std::string& checkSum, uint64_t offset,
+                      uint64_t actual_size);
     std::string serialize();
     void deserialize(const std::string& data);
     std::string to_string(bool isFull = false) const;
